Move ServiceScenario to a dedicated module so androidtest does not need to include Robolectric
