+++ audiorecorder/build.gradle
@@ -54,6 +54,7 @@ dependencies {
 
     testImplementation project(path: ':androidtest')
     testImplementation project(path: ':testshared')
+    testImplementation project(path: ':servicetest')
     testImplementation Dependencies.junit
     testImplementation Dependencies.robolectric
     testImplementation Dependencies.androidx_test_ext_junit
