+++ collect_app/build.gradle
@@ -390,9 +390,7 @@ dependencies {
     testImplementation Dependencies.androidx_test_core_ktx
     testImplementation Dependencies.kotlinx_coroutines_test
 
-    androidTestImplementation(project(':androidtest')) {
-        exclude group: 'org.robolectric' // Robolectric breaks things when running on device
-    }
+    androidTestImplementation project(':androidtest')
 
     // Used to generate documentation screenshots.
     androidTestImplementation Dependencies.fastlane_screengrab
