+++ servicetest/README.md
@@ -1,0 +1,3 @@
+# servicetest
+
+Includes helpers (arguably missing from `androidx.test:core`) for writing local tests.
