+++ settings.gradle
@@ -23,3 +23,4 @@ include ':imageloader'
 include ':glide'
 include ':androidtest'
 include ':settings'
+include ':servicetest'
