+++ testshared/build.gradle
@@ -38,6 +38,7 @@ dependencies {
     implementation project(path: ':shared')
     implementation project(path: ':androidshared')
     implementation project(path: ':androidtest')
+    implementation project(path: ':servicetest')
     implementation Dependencies.androidx_recyclerview
     implementation Dependencies.kotlin_stdlib
     implementation Dependencies.androidx_core_ktx
